Performance improvement

+++ src/framebuffer/framebuffer_builder.cpp
@@ -28,7 +28,7 @@ FramebufferBuilder& FramebufferBuilder::withTexture(GLuint attachment, const glm
 	glGenTextures(1, &textureId);
 
 	glBindTexture(GL_TEXTURE_2D, textureId);
-	glTexImage2D(GL_TEXTURE_2D, 0, GL_RGB32F, dimensions.x, dimensions.y, 0, GL_RGB, GL_FLOAT, nullptr);
+	glTexImage2D(GL_TEXTURE_2D, 0, GL_RGB16F, dimensions.x, dimensions.y, 0, GL_RGB, GL_FLOAT, nullptr);
 	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
 	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
 	_framebuffer->addTextureAttachment(attachment, textureId);
